utils/monitor-deployment/requirements.txt
psycopg2-binary==2.9.4
prometheus-client==0.15.0
structlog==22.3.0
